+++ lib/hashtable.hh
@@ -87,7 +87,7 @@ namespace khmer {
     // count every k-mer in the FASTA file.
     void consume_fasta(const std::string &filename,
 		       unsigned int &total_reads,
-		       unsigned int &n_consumed,
+		       unsigned long long &n_consumed,
 		       HashIntoType lower_bound = 0,
 		       HashIntoType upper_bound = 0,
 		       ReadMaskTable ** readmask = NULL,
